requirement.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
pycryptodome==3.20.0
